+++ incubator/lint-json/src/index.ts
@@ -1,0 +1,11 @@
+export { compareValues } from "./compare.ts";
+export { getJSONPathSegments, isJSONObject } from "./utilities.ts";
+export { createJSONValidator } from "./validator.ts";
+
+export type {
+  JSONObject,
+  JSONValidator,
+  JSONValidatorOptions,
+  JSONValue,
+  JSONValuePath,
+} from "./types.ts";