+++ incubator/lint-lockfile/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @rnx-kit/lint-lockfile
 
+## 0.1.4
+
+### Patch Changes
+
+- Updated dependencies [f53f04d]
+  - @rnx-kit/config@0.8.0
+
 ## 0.1.3
 
 ### Patch Changes