+++ packages/align-deps/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log - @rnx-kit/align-deps
 
+## 3.5.0
+
+### Minor Changes
+
+- 61a2965: Bump package to bundle updates to config with minor version changes
+
+## 3.4.8
+
+### Patch Changes
+
+- 3fa3488: Bumped `react-native-test-app` to support React Native for macOS/Windows up to 0.82
+
 ## 3.4.7
 
 ### Patch Changes